Add padding in sprite manager to counter artifacts
We currently use SDL_RenderSetLogicalSize to scale 640x480 game area
to arbitrarily sized window, however it does not provide proportional
scaling for every pixel. Because of that, pixels from adjacent sprites
may be renderer. Use padding to counter that. Later, custom scaler may
be implemented (based on coordinate offsets for SDL2pp::Renderer::Copy
or by rendering to a 640x480 target and then scaling it to the screen)